fix:一括配置で必要な処理をRuntime側へ移行
