+++ PlateauToolkit.Sandbox/Editor/PlateauSandboxBulkPlaceHierarchyView.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using PlateauToolkit.Sandbox.Runtime;
+using System.Collections.Generic;
 using UnityEditor.IMGUI.Controls;
 using UnityEngine;
 using UnityEngine.Events;
@@ -31,16 +32,6 @@ namespace PlateauToolkit.Sandbox.Editor
         }
 
         public MultiColumnHeaderState.Column[] Columns { get; }
-    }
-
-    class PlateauSandboxBulkPlaceHierarchyItem
-    {
-        public int ID { get; set; }
-        public string CategoryName { get; set; }
-        public int Count { get; set; }
-        public string PrefabName { get; set; } = string.Empty;
-        public int PrefabConstantID { get; set; } = -1;
-        public UnityEvent OnClicked { get; private set; } = new UnityEvent();
     }
 
     class PlateauSandboxBulkPlaceHierarchyViewItem : TreeViewItem
