+++ PlateauToolkit.Sandbox/Runtime/PlateauToolkit.Sandbox.asmdef
@@ -5,7 +5,9 @@
         "Unity.InputSystem",
         "Unity.Mathematics",
         "Unity.Splines",
-        "PlateauToolkit.Foundation"
+        "PlateauToolkit.Foundation",
+        "PLATEAU.Runtime",
+        "libplateau"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
